test(cmd): fold runModule tests into generator_test.go, drop misleading 'e2e' name
The TestRunModule_* tests are tag-less, mock-based, in-process unit tests that
run in CI (no terraform/cloud) — the 'e2e' filename implied the repo's
//go:build integration suite under test/integration, which they are not. Fold
them into the canonical cmd unit-test file, rename the helper, and reword doc
comments to say 'in-process'. Pure relocation; no behavior change.

Coverage from the removed generator_neo4j_test.go remains in CI: the content
gate is covered by TestNeo4jSink_* (pkg/plugin) and the runModule routing
decision by these TestRunModule_* tests.
